# Migrated code to not clear the view container on every update