--- a/src/app/reexportr-src/reexportr.directive.ts
+++ b/src/app/reexportr-src/reexportr.directive.ts
@@ -11,7 +11,7 @@ import {
 @Directive({
   selector: "[reExportr]"
 })
-export class ReExportrDirective implements OnInit, OnDestroy {
+export class ReExportrDirective implements OnInit {
   context = {
     $implicit: undefined,
     reExportr: undefined
@@ -24,16 +24,9 @@ export class ReExportrDirective implements OnInit, OnDestroy {
     this.embeddedView = this._vcr.createEmbeddedView(this._tr, this.context);
   }
 
-  ngOnDestroy() {
-    this._vcr.clear();
-  }
-
   @Input()
   set reExportr(val: any) {
     this.context.reExportr = val;
     this.context.$implicit = val;
-    if (this.embeddedView) {
-      this.embeddedView.detectChanges();
-    }
   }
 }