Looking at the code from this directive (as well as https://www.npmjs.com/package/ngx-let-directive and a few other packages that do similar to this), I quickly noticed that on every page render, it is clearing the entire view container and re-rendering the entire thing, regardless of if all of the data there-within requires changes or not. This breaks the pattern of change detection that Angular has built in. I figured I'd make a PR to fix this and utilize the internal change detection to handle change handling

The result is a 10x speedup in one of the small examples I threw together:

```typescript
import { Component } from '@angular/core';

@Component({
  selector: 'my-app',
  template: `
  <input placeholder="Change the value" [(ngModel)]="input"/>

<ng-container *reExportr="(input | uppercase) as upper">
	<h1>{{upper}}</h1>
  <p *ngFor="let val of array">This is a long counting message, there should be a lot of them {{val}}</p>
</ng-container>
  `
})
export class AppComponent  {
  input = '';

  array = Array.from(new Array(10000), (_, i) => i)
}

```

In this example, rendering a LOT of content on-screen that doesn't rely on the changed value.

Before:
https://stackblitz.com/edit/angularnglet-old

![A performance benchmark showing 1139ms in rendering time and 2675ms of scripting time](https://user-images.githubusercontent.com/9100169/70367915-f03e6b80-1859-11ea-9d17-8aec428d471f.png)

After:

https://stackblitz.com/edit/angularnglet

![A performance benchmark showing 131ms in rendering time and 268ms of scripting time](https://user-images.githubusercontent.com/9100169/70367923-0ba97680-185a-11ea-892b-18c4ddbfebf4.png)

In this (admittedly extreme) example, we're seeing a ~10x speedup