@rarmatei 2019-12-07
Amazing, thank you so much for creating this!!

Much more than the performance problem you illustrated, this directive used to completely destroy and re-create any inner components on each change of the input. 

I left just a few small comments, but otherwise looks good!